+++ src/main/java/org/team6/coffeebeanery/order/model/OrderDetail.java
@@ -11,11 +11,11 @@ import org.team6.coffeebeanery.product.model.Product;
 public class OrderDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderDetailId; //주문 상세 ID
+    private Long orderDetailId; //주문 상세 ID
     
     private Integer productQuantity; //주문 상품 수량
     
-    private Integer orderPrice; //주문 시점 가격
+    private Long orderPrice; //주문 시점 가격
     
     @OneToOne
     private Product product; //주문 상품
