+++ src/main/java/org/team6/coffeebeanery/order/repository/OrderRepository.java
@@ -1,7 +1,7 @@
 package org.team6.coffeebeanery.order.repository;
 
-import org.team6.coffeebeanery.order.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.team6.coffeebeanery.order.model.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
